# docs/swagger.yaml
basePath: /api/v1
definitions:
  swagmodel.ChPwd:
    properties:
      password:
        type: string
    type: object
  swagmodel.GetArticle:
    properties:
      body:
        type: string
      creatorID:
        type: integer
      id:
        type: integer
      imageURL:
        type: string
      title:
        type: string
    type: object
  swagmodel.GetComment:
    properties:
      articleID:
        type: integer
      body:
        type: string
      creatorID:
        type: integer
      id:
        type: integer
    type: object
...
  swagmodel.InputArticle:
    properties:
      body:
        type: string
      imageURL:
        type: string
      title:
        type: string
    type: object
  swagmodel.InputComment:
    properties:
      body:
        type: string
    type: object
...
  swagmodel.LoginPayload:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  swagmodel.Register:
    properties:
      avatarURL:
        type: string
      bio:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  swagmodel.Response:
    properties:
      msg:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: hdkef11@gmail.com
    name: Hadekha Erfadila Fitra
  description: Bloggo is a REST Api for blogging
  title: Bloggo API
  version: "1.0"
paths:
  /articles:
    get:
      description: get all articles navigated by last-id and limit.
      parameters:
      - description: ID of the last article in recent array of article
        in: query
        name: last-id
        required: true
        type: integer
      - description: how many article you want to take
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/swagmodel.GetArticle'
            type: array
      summary: get all articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: create an article
      parameters:
      - description: create an article
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputArticle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: create an article
      tags:
      - articles
  /articles/{id}:
    delete:
      consumes:
      - application/json
      description: delete an article
      parameters:
      - description: ID of article
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: delete an article
      tags:
      - articles
    get:
      description: get detail of article specified by id
      parameters:
      - description: ID of article
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.GetArticle'
      summary: get one article
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: update an article
      parameters:
      - description: ID of article
        in: path
        name: id
        required: true
        type: integer
      - description: update an article
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputArticle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: update an article
      tags:
      - articles
  /ch-pwd:
    post:
      consumes:
      - application/json
      description: send authorization header with password payload to change password
      parameters:
      - description: change password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/swagmodel.ChPwd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: change user password
      tags:
      - auth
  /comments/{article-id}:
    get:
      description: get all comments navigated by last-id and limit.
      parameters:
      - description: ID of article
        in: path
        name: article-id
        required: true
        type: integer
      - description: ID of the last comment in recent array of comment
        in: query
        name: last-id
        required: true
        type: integer
      - description: how many comment you want to take
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/swagmodel.GetComment'
            type: array
      summary: get all comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: create an comment
      parameters:
      - description: ID of article
        in: path
        name: article-id
        required: true
        type: integer
      - description: create an comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: create an comment
      tags:
      - comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: delete an comment
      parameters:
      - description: ID of comment
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: delete an comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: update an comment
      parameters:
      - description: ID of comment
        in: path
        name: id
        required: true
        type: integer
      - description: update an comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: update an comment
      tags:
      - comments
...
  /login:
    post:
      consumes:
      - application/json
      description: send username and password to get JWT Token via cookie
      parameters:
      - description: create an user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/swagmodel.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      summary: send username and password to get JWT Token
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      description: send username, password, bio, and avatar url
      parameters:
      - description: create a regular user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/swagmodel.Register'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      summary: register a regular user
      tags:
      - auth
...
securityDefinitions:
  AuthToken:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"
...